# apps/web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PairRequest": {
        "properties": {
          "agent_version": {
            "maxLength": 50,
            "title": "Agent Version",
            "type": "string"
          },
          "company_guid": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Company Guid",
            "type": "string"
          },
          "company_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Company Name",
            "type": "string"
          },
          "device_name": {
            "maxLength": 255,
            "minLength": 2,
            "title": "Device Name",
            "type": "string"
          },
          "license_mode": {
            "default": "",
            "maxLength": 50,
            "title": "License Mode",
            "type": "string"
          },
          "pairing_code": {
            "maxLength": 200,
            "minLength": 20,
            "title": "Pairing Code",
            "type": "string"
          },
          "protocols": {
            "items": {
              "type": "string"
            },
            "maxItems": 4,
            "title": "Protocols",
            "type": "array"
          },
          "public_key": {
            "maxLength": 100,
            "minLength": 40,
            "title": "Public Key",
            "type": "string"
          },
          "tally_version": {
            "default": "",
            "maxLength": 100,
            "title": "Tally Version",
            "type": "string"
          }
        },
        "required": [
          "pairing_code",
          "device_name",
          "company_name",
          "company_guid",
          "agent_version",
          "public_key"
        ],
        "title": "PairRequest",
        "type": "object"
      },
      "PairingCodeRequest": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          }
        },
        "required": [
          "client_id"
        ],
        "title": "PairingCodeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "firmOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tally-agent/devices": {
      "get": {
        "description": "Return device health without ever exposing token digests.",
        "operationId": "list_devices_api_tally_agent_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Devices",
        "tags": [
          "tally-agent"
        ]
      }
    },
    "/api/tally-agent/disconnect": {
      "post": {
        "operationId": "disconnect_device_api_tally_agent_disconnect_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FirmOS-Device",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 36,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Device"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Firm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Firm"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 20,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Timestamp"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Nonce",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Nonce"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Body-SHA256",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Body-Sha256"
            }
          },
          {
            "in": "header",
            "name": "X-FirmOS-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Firmos-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disconnect Device",
        "tags": [
          "tally-agent"
        ]
      }
    },
...
    "/api/tally-agent/pair": {
      "post": {
        "operationId": "pair_device_api_tally_agent_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair Device",
        "tags": [
          "tally-agent"
        ]
      }
    },
    "/api/tally-agent/pairing-code": {
      "post": {
        "operationId": "create_pairing_code_api_tally_agent_pairing_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pairing Code",
        "tags": [
          "tally-agent"
        ]
      }
    },
...
  }
}